commit-homelocal

+++ Mbc5/Forms/MemoryBook/frmSales.cs
@@ -4422,7 +4422,7 @@ namespace Mbc5.Forms.MemoryBook
                 {
                     //getinvoices
                     var sqlquery = new SQLCustomClient();
-                    var cmdtext = @"Select X.SalesInvno AS SalesInvoice,I.Invno AS Invoice,CAST(I.qtedate AS date)AS InvoiceDate,I.nocopies AS Quantity,I.InvTot As Total From Xtra X 
+                    var cmdtext = @"Select X.SalesInvno AS SalesInvoice, I.Invno AS Invoice,CAST(I.qtedate AS date)AS InvoiceDate,I.nocopies AS Quantity,I.InvTot As Total From Xtra X 
                                     Left Join Sales_ExtraInvoice SX ON X.SalesInvno=SX.SalesInvoice
                                      Left Join Invoice I ON SX.XtraInvoice=I.Invno 
                                         Where SX.SalesInvoice=@Invno";
@@ -4740,7 +4740,7 @@ namespace Mbc5.Forms.MemoryBook
         public decimal SalesInvoice { get; set; }
         public int Invoice { get; set; }
         public DateTime InvoiceDate { get; set; }
-        public decimal Quantity { get; set; }
+        public int Quantity { get; set; }
         public decimal Total { get; set; }
     }
     public class XtraInvoicePrint
